[compiler] Refactor traverseFunction in Entrypoint
Previously, we would traverse and compile a function and add the
compiled function to an array of compiled functions captured by context.

With this PR, rather than modifying the captured array, the compiled
result is returned. A small helper function is used to wrap this call
and update the array instead.

This change will let us reuse this traversal and compilation pipeline
for newly created functions (like when we outline jsx).

ghstack-source-id: 13848e6dea29f05ad91850e5654d616fc5108d01
Pull Request resolved: https://github.com/facebook/react/pull/30084