[compiler] Introduce OutlinedFunctionExpression HIR
This is very similar to the existing FunctionExpression HIR node
intentionally, so that we can share all the infra between them.

There's a new `outlinedFunc` field on this that will hold the newly
created function that needs to be compiled outside the function body.

The `loweredFunc` will hold the updated HIR for rendering the
outlinedFunc in the callback.

ghstack-source-id: 06a92eb575290c086e128513e4c4d0f96194be2f
Pull Request resolved: https://github.com/facebook/react/pull/30088